@@ -12,18 +12,18 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^19.2.14",
+    "@angular/animations": "^20.3.15",
     "@angular/cdk": "^19.2.14",
-    "@angular/common": "^19.2.14",
-    "@angular/compiler": "^19.2.14",
-    "@angular/core": "^19.2.14",
-    "@angular/elements": "^19.2.14",
-    "@angular/forms": "^19.2.14",
+    "@angular/common": "^20.3.15",
+    "@angular/compiler": "^20.3.15",
+    "@angular/core": "^20.3.15",
+    "@angular/elements": "^20.3.15",
+    "@angular/forms": "^20.3.15",
     "@angular/material": "^19.2.14",
     "@angular/material-luxon-adapter": "^19.2.14",
-    "@angular/platform-browser": "^19.2.14",
-    "@angular/platform-browser-dynamic": "^19.2.14",
-    "@angular/router": "^19.2.14",
+    "@angular/platform-browser": "^20.3.15",
+    "@angular/platform-browser-dynamic": "^20.3.15",
+    "@angular/router": "^20.3.15",
     "@microsoft/signalr": "^8.0.7",
     "@mr-samani/event-bus": "^1.0.2",
     "@ngx-translate/core": "^16.0.4",
@@ -54,9 +54,9 @@
     "zone.js": "~0.15.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^19.2.1",
-    "@angular/cli": "^19.2.1",
-    "@angular/compiler-cli": "^19.2.0",
+    "@angular/build": "^20.3.13",
+    "@angular/cli": "^20.3.13",
+    "@angular/compiler-cli": "^20.3.15",
     "@openapitools/openapi-generator-cli": "^2.21.3",
     "@types/jasmine": "~5.1.0",
     "@types/lodash-es": "^4.17.12",
@@ -69,6 +69,6 @@
     "karma-jasmine-html-reporter": "~2.1.0",
     "prettier": "^3.5.3",
     "ts-node": "^10.9.2",
-    "typescript": "~5.7.2"
+    "typescript": "~5.9.3"
   }
 }
