@@ -1,4 +1,4 @@
-import { CommonModule } from '@angular/common';
+
 import { Component, Injector, OnInit } from '@angular/core';
 import { AppComponentBase } from '@app/app-component-base';
 import { TranslateModule } from '@ngx-translate/core';
@@ -9,7 +9,7 @@ import { SharedModule } from '@shared/shared.module';
   templateUrl: './user-profile.component.html',
   styleUrls: ['./user-profile.component.scss'],
   standalone: true,
-  imports: [CommonModule, SharedModule],
+  imports: [SharedModule],
 })
 export class UserProfileComponent extends AppComponentBase implements OnInit {
   constructor(injector: Injector) {
